requirements.txt
RPi.GPIO
psutil
apscheduler
